MontageManager added and all table models are now sortable.
Plus various small changes and fixes.


Former-commit-id: a12ca4995bc9ff75be442f69b4b1015ad43e7ff5
Former-commit-id: a36ff49ea26c2678956227fdf19d26ce1858d47a

+++ App/src/canvas.h
@@ -67,7 +67,7 @@ private:
 	}
 	void drawBlock(const SignalBlock& block, const std::vector<std::tuple<int, int, int, int>>& singleChannelEvents);
 	void setUniformChannel(GLuint program, int channel, const SignalBlock& block);
-	void setUniformColor(GLuint program, const QColor& color, float opacity);
+	void setUniformColor(GLuint program, const QColor& color, double opacity);
 	void checkGLMessages()
 	{
 		for (const auto& m : log()->loggedMessages())
